pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wispr-local-dictation"
version = "0.1.0"
description = "Local macOS dictation daemon powered by whisper.cpp and Ollama."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pyobjc-framework-Cocoa>=10.3",
  "pyobjc-framework-Quartz>=10.3",
  "sounddevice>=0.4.7",
]

[project.scripts]
wispr = "wispr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

